refactor(web): split api.ts into domain-specific modules [H4]
Split the monolithic 1,556-line api.ts into 13 domain-specific modules
under web/src/lib/api/:

- client.ts: base ApiClient class, HTTP methods, CSRF, auth helpers
- auth.ts: login, register, logout, password reset
- content.ts: content CRUD, upload, folders
- displays.ts: display management, pairing, groups, QR overlay
- playlists.ts: playlist CRUD, items, reorder
- schedules.ts: schedule CRUD, conflict checking
- organizations.ts: org settings
- analytics.ts: analytics, device uptime
- users.ts: team management, audit logs
- notifications.ts: notification CRUD
- billing.ts: subscription, plans, API keys
- admin.ts: admin operations (plans, orgs, users, config, security)
- templates.ts: template library, AI generation
- widgets.ts: widgets and layouts
- support.ts: support requests and messages

Uses TypeScript declaration merging to extend ApiClient prototype
in each domain file. Barrel index.ts re-exports everything.
Original api.ts becomes a thin re-export for backwards compatibility.
